Split service file into managed service and request handler files
